+++ src/plone/restapi/tests/test_workflow.py
@@ -5,6 +5,7 @@ from base64 import b64encode
 from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.app.testing import login
+from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.testing import PLONE_RESTAPI_DX_INTEGRATION_TESTING
 from unittest import TestCase
 from zExceptions import NotFound
@@ -42,6 +43,22 @@ class TestWorkflowInfo(TestCase):
         self.assertIn('transitions', info)
         transitions = info['transitions']
         self.assertEqual(2, len(transitions))
+
+    def test_collapsed_workflow_info_in_content_serialization(self):
+        serializer = getMultiAdapter((self.doc1, self.request),
+                                     ISerializeToJson)
+        obj = serializer()
+        self.assertIn('workflow', obj['@components'])
+        self.assertIn('@id', obj['@components']['workflow'])
+
+    def test_expanded_workflow_info_in_content_serialization(self):
+        self.request.form.update({'expand': 'workflow'})
+        serializer = getMultiAdapter((self.doc1, self.request),
+                                     ISerializeToJson)
+        obj = serializer()
+        self.assertIn('workflow', obj['@components'])
+        self.assertIn('transitions', obj['@components']['workflow'])
+        self.assertIn('history', obj['@components']['workflow'])
 
 
 class TestWorkflowTransition(TestCase):